fix: dto mapper 함수의 위치를 서비스 계층이 private 함수로 이동
